+++ ethereumkit/src/main/java/io/horizontalsystems/ethereumkit/spv/net/les/messages/AnnounceMessage.kt
@@ -2,6 +2,7 @@ package io.horizontalsystems.ethereumkit.spv.net.les.messages
 
 import io.horizontalsystems.ethereumkit.core.toHexString
 import io.horizontalsystems.ethereumkit.spv.core.toBigInteger
+import io.horizontalsystems.ethereumkit.spv.core.toLong
 import io.horizontalsystems.ethereumkit.spv.net.IInMessage
 import io.horizontalsystems.ethereumkit.spv.rlp.RLP
 import io.horizontalsystems.ethereumkit.spv.rlp.RLPList
@@ -10,16 +11,16 @@ import java.math.BigInteger
 class AnnounceMessage(payload: ByteArray) : IInMessage {
 
     val blockHash: ByteArray
-    val blockHeight: BigInteger
+    val blockHeight: Long
     val blockTotalDifficulty: BigInteger
-    val reorganizationDepth: BigInteger
+    val reorganizationDepth: Long
 
     init {
         val params = RLP.decode2(payload)[0] as RLPList
         blockHash = params[0].rlpData ?: byteArrayOf()
-        blockHeight = params[1].rlpData.toBigInteger()
+        blockHeight = params[1].rlpData.toLong()
         blockTotalDifficulty = params[2].rlpData.toBigInteger()
-        reorganizationDepth = params[3].rlpData.toBigInteger()
+        reorganizationDepth = params[3].rlpData.toLong()
     }
 
     override fun toString(): String {
